Pet_industry_prediction_streamlit/requirements.txt
numpy
pandas
streamlit
seaborn
matplotlib
koreanize_matplotlib
wordcloud
pip>=22.3.1
pillow
tableauserverclient